+++ integration_tests/models/groups_matrix_regression_chol_unoptimized.sql
@@ -1,0 +1,17 @@
+{{
+  config(
+    materialized="table"
+  )
+}}
+select * from {{
+  dbt_linreg.ols(
+    table=ref('groups_matrix'),
+    endog='y',
+    exog=['x1', 'x2', 'x3'],
+    group_by=['gb_var'],
+    format='long',
+    method='chol',
+    method_options={'subquery_optimization': False}
+  )
+}}
+order by gb_var, variable_name