+++ integration_tests/models/simple_0var_regression_long_fwl.sql
@@ -1,0 +1,14 @@
+{{
+  config(
+    materialized="table"
+  )
+}}
+select * from {{
+  dbt_linreg.ols(
+    table=ref('simple_matrix'),
+    endog='y',
+    exog=[],
+    format='long',
+    format_options={'round': 5}
+  )
+}}