+++ integration_tests/models/simple_1var_regression_long_fwl.sql
@@ -5,9 +5,10 @@
 }}
 select * from {{
   dbt_linreg.ols(
-    table=ref('collinear_matrix'),
+    table=ref('simple_matrix'),
     endog='y',
-    exog=['x1', 'x2', 'x3', 'x4', 'x5'],
-    format='long'
+    exog=['xa'],
+    format='long',
+    format_options={'round': 5}
   )
 }}