+++ integration_tests/tests/test_collinear_matrix_regression_fwl.sql
@@ -17,9 +17,10 @@ expected as (
 )
 
 select base.variable_name
-from {{ ref('collinear_matrix_regression') }} as base
+from {{ ref('collinear_matrix_regression_fwl') }} as base
 full outer join expected
 on base.variable_name = expected.variable_name
 where
-  round(base.coefficient, 7) - round(expected.coefficient, 7)
+  round(base.coefficient, 7) != round(expected.coefficient, 7)
   or base.coefficient is null
+  or expected.coefficient is null