+++ integration_tests/tests/test_groups_matrix_regression_fwl.sql
@@ -24,11 +24,12 @@ expected as (
 )
 
 select base.variable_name
-from {{ ref('groups_matrix_regression') }} as base
+from {{ ref('groups_matrix_regression_fwl') }} as base
 full outer join expected
 on
   base.gb_var = expected.gb_var
   and base.variable_name = expected.variable_name
 where
-  round(base.coefficient, 7) - round(expected.coefficient, 7)
+  round(base.coefficient, 7) != round(expected.coefficient, 7)
   or base.coefficient is null
+  or expected.coefficient is null