+++ integration_tests/tests/test_simple_0var_regression_wide.sql
@@ -7,7 +7,7 @@ expected as (
 )
 
 select base.*
-from {{ ref('simple_2var_regression_wide') }} as base, expected
+from {{ ref('simple_0var_regression_wide') }} as base, expected
 where not (
   base.const = expected.const
 )