+++ scripts.py
@@ -1,4 +1,5 @@
 import os.path as op
+import warnings
 from typing import NamedTuple
 from typing import Optional
 from typing import Protocol
@@ -8,6 +9,13 @@ import pandas as pd
 import rich_click as click
 import statsmodels.api as sm
 from tabulate import tabulate
+
+
+# Suppress iteritems warning
+warnings.simplefilter("ignore", category=FutureWarning)
+
+# No scientific notation
+np.set_printoptions(suppress=True)
 
 
 DIR = op.dirname(__file__)
@@ -185,6 +193,11 @@ def cli():
 
 @cli.command("regress")
 @click.argument("table")
+@click.option("--const/--no-const",
+              default=True,
+              type=click.BOOL,
+              show_default=True,
+              help="If true, add the constant term.")
 @click.option("--columns", "-c",
               default=None,
               type=click.INT,
@@ -197,16 +210,16 @@ def cli():
               help="Alpha for the regression.")
 @click_option_size()
 @click_option_seed()
-def regress(table: str, columns: int, alpha: float, size: int, seed: int):
+def regress(table: str, const: bool, columns: int, alpha: float, size: int, seed: int):
     callback = ALL_TEST_CASES[table]
 
-    click.echo(click.style("=" * 80, fg="red"))
+    click.echo(click.style("=" * 80, fg="blue"))
     click.echo(
-        click.style("Test case: ", fg="red", bold=True)
+        click.style("Test case: ", fg="blue", bold=True)
         +
-        click.style(table, fg="red")
+        click.style(table, fg="blue")
     )
-    click.echo(click.style("=" * 80, fg="red"))
+    click.echo(click.style("=" * 80, fg="blue"))
 
     test_case = callback(size, seed)
 
@@ -214,7 +227,10 @@ def regress(table: str, columns: int, alpha: float, size: int, seed: int):
         x_cols = test_case.x_cols
     else:
         # K plus Constant (1)
-        x_cols = test_case.x_cols[:columns]
+        x_cols = test_case.x_cols[:columns+1]
+
+    if not const:
+        x_cols = [i for i in x_cols if i != "const"]
 
     def _run_model(cond=None):
         if cond is None:
@@ -222,20 +238,24 @@ def regress(table: str, columns: int, alpha: float, size: int, seed: int):
         y = test_case.df.loc[cond, test_case.y_col]
         x_mat = test_case.df.loc[cond, x_cols]
         if alpha:
-            alpha_arr = [0, *([alpha] * (len(x_mat.columns) - 1))]
+            if const:
+                alpha_arr = [0, *([alpha] * (len(x_mat.columns) - 1))]
+            else:
+                alpha_arr = [alpha] * len(x_mat.columns)
             model = sm.OLS(
                 y,
                 x_mat
             ).fit_regularized(L1_wt=0, alpha=alpha_arr)
         else:
             model = sm.OLS(y, x_mat).fit()
+        res_df = pd.DataFrame(index=x_mat.columns)
+        res_df["coef"] = model.params
+        res_df["stderr"] = model.bse
+        res_df["tstat"] = res_df["coef"] / res_df["stderr"]
         click.echo(
             tabulate(
-                pd.DataFrame(
-                    {"coefficient": model.params},
-                    index=x_mat.columns
-                ),
-                headers=["coefficient", "value"],
+                res_df,
+                headers=["column name", "coef", "stderr", "tstat"],
                 disable_numparse=True,
                 tablefmt="psql",
             )